emorag_core/requirements.txt
sentence-transformers
torch
numpy
scikit-learn
tqdm
pillow
